backend/ai/pyproject.toml
[project]
name = "indestructibleeco-ai"
version = "1.0.0"
description = "AI Engine Service - Folding, Compute, Index, Service Engines"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "httpx>=0.27.0",
    "redis>=5.0.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.5.0",
    "faiss-cpu>=1.8.0",
    "neo4j>=5.20.0",
    "elasticsearch[async]>=8.15.0",
    "grpcio>=1.65.0",
    "grpcio-tools>=1.65.0",
    "strawberry-graphql[fastapi]>=0.235.0",
    "aiokafka>=0.11.0",
    "prometheus-client>=0.21.0",
    "gunicorn>=22.0.0",
    "sentence-transformers>=3.0.0",
    "networkx>=3.3",
    "scipy>=1.14.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.5.0",
    "mypy>=1.11.0",
]

[build-system]
requires = ["setuptools>=70.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true